+++ tests/test_communication.py
@@ -6,11 +6,21 @@ import pytest
 from replit_river.client import Client
 from replit_river.error_schema import RiverError
 from replit_river.transport_options import MAX_MESSAGE_BUFFER_SIZE
-from tests.conftest import common_handlers, deserialize_error, deserialize_response, serialize_request
+from tests.common_handlers import (
+    basic_rpc_method,
+    basic_stream,
+    basic_subscription,
+    basic_upload,
+)
+from tests.conftest import (
+    deserialize_error,
+    deserialize_response,
+    serialize_request,
+)
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("handlers", [common_handlers])
+@pytest.mark.parametrize("handlers", [{**basic_rpc_method}])
 async def test_rpc_method(client: Client) -> None:
     response = await client.send_rpc(
         "test_service",
@@ -24,7 +34,7 @@ async def test_rpc_method(client: Client) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("handlers", [common_handlers])
+@pytest.mark.parametrize("handlers", [{**basic_upload}])
 async def test_upload_method(client: Client) -> None:
     async def upload_data() -> AsyncGenerator[str, None]:
         yield "Data 1"
@@ -45,7 +55,7 @@ async def test_upload_method(client: Client) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("handlers", [common_handlers])
+@pytest.mark.parametrize("handlers", [{**basic_upload}])
 async def test_upload_more_than_send_buffer_max(client: Client) -> None:
     iterations = MAX_MESSAGE_BUFFER_SIZE * 2
 
@@ -67,7 +77,7 @@ async def test_upload_more_than_send_buffer_max(client: Client) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("handlers", [common_handlers])
+@pytest.mark.parametrize("handlers", [{**basic_upload}])
 async def test_upload_empty(client: Client) -> None:
     async def upload_data(enabled: bool = False) -> AsyncGenerator[str, None]:
         if enabled:
@@ -87,7 +97,7 @@ async def test_upload_empty(client: Client) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("handlers", [common_handlers])
+@pytest.mark.parametrize("handlers", [{**basic_subscription}])
 async def test_subscription_method(client: Client) -> None:
     async for response in client.send_subscription(
         "test_service",
@@ -102,7 +112,7 @@ async def test_subscription_method(client: Client) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("handlers", [common_handlers])
+@pytest.mark.parametrize("handlers", [{**basic_stream}])
 async def test_stream_method(client: Client) -> None:
     async def stream_data() -> AsyncGenerator[str, None]:
         yield "Stream 1"
@@ -131,7 +141,7 @@ async def test_stream_method(client: Client) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("handlers", [common_handlers])
+@pytest.mark.parametrize("handlers", [{**basic_stream}])
 async def test_stream_empty(client: Client) -> None:
     async def stream_data(enabled: bool = False) -> AsyncGenerator[str, None]:
         if enabled:
@@ -154,7 +164,7 @@ async def test_stream_empty(client: Client) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("handlers", [common_handlers])
+@pytest.mark.parametrize("handlers", [{**basic_upload, **basic_stream}])
 async def test_multiplexing(client: Client) -> None:
     async def upload_data() -> AsyncGenerator[str, None]:
         yield "Upload Data 1"
