+++ tests/test_handshake.py
@@ -7,7 +7,6 @@ from websockets.server import serve
 
 from replit_river.server import Server
 from replit_river.transport_options import TransportOptions
-from tests.conftest import common_handlers
 
 
 @pytest.fixture
@@ -16,7 +15,7 @@ def transport_options() -> TransportOptions:
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("handlers", [common_handlers])
+@pytest.mark.parametrize("handlers", [{}])
 async def test_handshake_timeout(server: Server) -> None:
     async with serve(server.serve, "localhost", 8765):
         start = time()
